+++ ext_localconf.php
@@ -1,4 +1,18 @@
 <?php
 defined('TYPO3_MODE') or die();
 
-$GLOBALS['TYPO3_CONF_VARS']['SYS']['fluid']['namespaces']['schema'][] = 'Brotkrueml\\SchemaPending\\ViewHelpers';
+(function() {
+    if (!(new Brotkrueml\Schema\Compatibility\Compatibility)->isPsr14EventDispatcherAvailable()) {
+        $signalSlotDispatcher = TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(
+            TYPO3\CMS\Extbase\SignalSlot\Dispatcher::class
+        );
+        $signalSlotDispatcher->connect(
+            Brotkrueml\Schema\Core\Model\AbstractType::class,
+            'registerAdditionalTypeProperties',
+            Brotkrueml\SchemaPending\EventListener\RegisterAdditionalProperties::class,
+            '__invoke'
+        );
+    }
+
+    $GLOBALS['TYPO3_CONF_VARS']['SYS']['fluid']['namespaces']['schema'][] = 'Brotkrueml\\SchemaPending\\ViewHelpers';
+})();